pyproject.toml
[project]
name = "lpt-event"
dynamic = ["version"]
description = "LPT Event Management Application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
]

[dependency-groups]
dev = [
    "apx",
]

[tool.apx.metadata]
app-name = "lpt-event"
app-slug = "lpt_event"
app-module = "lpt_event.backend.app:app"
metadata-path = "src/lpt_event/_metadata.py"


[tool.uv.sources]
apx = { git = "https://github.com/databricks-solutions/apx.git" }

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
format-jinja = "{{version}}"

[tool.hatch.build.hooks.version]
path = "src/lpt_event/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/lpt_event/__dist__", "src/lpt_event/_metadata.py"]
exclude = ["src/lpt_event/ui"]